gcp/requirements.txt
tensorflow==2.8.0
google-cloud-storage==2.2.1
Pillow==8.4.0